pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "application-tracker"
version = "0.1.0"
description = "Self-hosted job application lifecycle tracker."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13,<2",
    "email-validator>=2,<3",
    "fastapi>=0.110,<1",
    "pydantic-settings>=2,<3",
    "pwdlib[argon2]>=0.2,<1",
    "sqlalchemy>=2,<3",
    "uvicorn[standard]>=0.27,<1",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27,<1",
    "pytest>=8,<9",
    "ruff>=0.6,<1",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
